+++ lib/Components/Cards/AbsenceCard.dart
@@ -1,11 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:rainbow_app/Backend/APIS/UserEmployeeService.dart';
-import 'package:rainbow_app/Pages/Calendar/AbsenceRequestPage/AbsenceRequestPage.dart';
-import 'package:rainbow_app/Pages/Calendar/AbsenceListPage/AbsenceListPage.dart';
-import 'package:rainbow_app/Theme/ThemeColor.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:rainbow_app/Backend/APIS/UserEmployeeService.dart';
+import 'package:rainbow_app/Pages/Calendar/AbsenceListPage/AbsenceListPage.dart';
+import 'package:rainbow_app/Pages/Calendar/AbsenceRequestPage/AbsenceRequestPage.dart';
+import 'package:rainbow_app/Theme/ThemeColor.dart';
 
-import '../../Backend/Models/UserModel.dart';
 import '../../Theme/ThemeTextStyle.dart';
 
 class AbsenceCard extends StatelessWidget {
@@ -14,7 +13,7 @@ class AbsenceCard extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     UserEmployeeService userEmployeeService = UserEmployeeService();
-    int leaveBalance = 0;
+    double leaveBalance = 0;
     userEmployeeService.getLeaveBalance().then((value) {
       leaveBalance = leaveBalance + value;
     });
