+++ lib/Pages/Calendar/AbsenceRequestPage/AbsenceRequestPage.dart
@@ -12,21 +12,16 @@ import 'package:intl/intl.dart';
 import 'package:open_file/open_file.dart';
 import 'package:rainbow_app/Backend/APIS/AbsenceService.dart';
 import 'package:rainbow_app/Backend/Models/Absence.dart';
-import 'package:rainbow_app/Backend/Models/UserModel.dart';
 import 'package:rainbow_app/Components/TextInputs/InputField.dart';
 import 'package:rainbow_app/Components/TextInputs/MultiLineInputField.dart';
-import 'package:rainbow_app/Components/Tiles/AbsenceTile.dart';
 
 import '../../../Backend/APIS/UserEmployeeService.dart';
 import '../../../Components/Buttons/Button.dart';
-import '../../../Components/Navigation.dart';
-import '../../../Components/TextInputs/DropdownTextField.dart';
 import '../../../Components/TextInputs/SmallInputField.dart';
 import '../../../Components/Tiles/Absence2Tile.dart';
 import '../../../Theme/ThemeColor.dart';
 import '../../../Theme/ThemeTextStyle.dart';
 import '../Absence/AbsenceInfoPage.dart';
-import '../AbsenceListPage/AbsenceListPage.dart';
 
 class AbsenceRequestPage extends StatefulWidget {
   const AbsenceRequestPage({Key? key}) : super(key: key);
@@ -51,7 +46,7 @@ class _AbsenceRequestPageState extends State<AbsenceRequestPage> {
   List<Absence> absences = [];
   int listLength = 0;
   String newValue = "";
-  int userLeaveBalance = 0;
+  double userLeaveBalance = 0;
   String fileName = "Choose file";
   List<String> items = [];
   late File image;
@@ -294,7 +289,7 @@ class _AbsenceRequestPageState extends State<AbsenceRequestPage> {
         ));
   }
 
-  void sendLeaveRequest(int userLeaveBalance, BuildContext context) {
+  void sendLeaveRequest(double userLeaveBalance, BuildContext context) {
     if (userLeaveBalance > getDateDifferenceInDays(fromDate, toDate)) {
       if (numberController.text.isNotEmpty) {
         AbsenceRequest absenceRequest = AbsenceRequest(
@@ -308,7 +303,7 @@ class _AbsenceRequestPageState extends State<AbsenceRequestPage> {
                       ? 1014
                       : 1003,
           days: getDateDifferenceInDays(fromDate, toDate),
-          uaAantaluren: int.parse(numberController.text),
+          uaAantaluren: double.parse(numberController.text),
           uaOpmerking: controller_3.text,
           fileBytes: fileAsString,
           typeFile: fileName,
@@ -631,7 +626,7 @@ class _AbsenceRequestPageState extends State<AbsenceRequestPage> {
     }
   }
 
-  String userLeaveHoursToDays(int userLeaveBalance) {
+  String userLeaveHoursToDays(double userLeaveBalance) {
     if (userLeaveBalance != 0) {
       double number = (userLeaveBalance / 8);
       return number.toString();
